+++ libs/react-components/community/src/components/table/components/table-layout/table-layout.tsx
@@ -3,7 +3,7 @@ import cn from 'classnames';
 import React from 'react';
 
 import { Col, Row } from '../../../../../../tedi/src/components/grid';
-import { useLabels } from '../../../../providers/label-provider';
+import { useLabels } from '../../../../../../tedi/src/providers/label-provider';
 import Button from '../../../button/button';
 import Placeholder from '../../../placeholder/placeholder';
 import Print from '../../../print/print';