+++ libs/react-components/tedi/src/index.ts
@@ -10,4 +10,5 @@ export * from './components/status-badge/status-badge';
 
 export * from './helpers';
 
+export * from './providers/label-provider';
 export * from './providers/style-provider/style-provider';